+++ backend-party-battle/ecosystem.config.cjs
@@ -1,6 +1,6 @@
 /* eslint-disable no-undef */
 /* eslint-disable @typescript-eslint/no-require-imports */
-const os = require('os');
+const os = require('os')
 
 /**
  * COLYSEUS CLOUD WARNING:
@@ -9,17 +9,18 @@ const os = require('os');
  */
 
 module.exports = {
-  apps: [{
-    name: "colyseus-app",
-    script: 'build/index.js',
-    time: true,
-    watch: false,
-    instances: os.cpus().length,
-    exec_mode: 'fork',
-    wait_ready: true,
-    env_production: {
-      NODE_ENV: 'production'
-    }
-  }],
-};
-
+  apps: [
+    {
+      name: 'colyseus-app',
+      script: 'build/index.js',
+      time: true,
+      watch: false,
+      instances: os.cpus().length,
+      exec_mode: 'fork',
+      wait_ready: true,
+      env_production: {
+        NODE_ENV: 'production',
+      },
+    },
+  ],
+}