+++ frontend-party-battle/app/index.tsx
@@ -6,8 +6,9 @@ import RainbowText from '@/components/rainbow-text'
 import { Button, ButtonText } from '@/components/ui/button'
 import { Heading } from '@/components/ui/heading'
 import { Input, InputField } from '@/components/ui/input'
-import { JoinRoomModal } from '@/components/ui/modal/join-room-modal'
 import { Text } from '@/components/ui/text'
+import useToastHelper from '@/components/ui/useToastHelper'
+import { JoinRoomModal } from '@/src/index/JoinRoomModal'
 
 import { usePlayerName } from '@/src/index/PlayerNameProvider'
 import { StoreBadges } from '@/src/index/StoreBadges'
@@ -21,6 +22,8 @@ export default function HomeScreen() {
   const { playerName, trimmedPlayerName, setPlayerName, isLoading: isLoadingPlayerName } = usePlayerName()
 
   const { createLobbyRoom, joinLobbyRoom, isLoading } = useLobbyRoomContext()
+  const [validationError, setValidationError] = useState<string | undefined>(undefined)
+  const { showError } = useToastHelper()
 
   const handleCreateRoom = async () => {
     await createLobbyRoom()
@@ -31,96 +34,108 @@ export default function HomeScreen() {
   }
 
   const handleJoinRoom = async (roomId: string) => {
-    await joinLobbyRoom(roomId)
+    const validationError = await joinLobbyRoom(roomId)
+
+    setShowJoinModal(false)
+
+    if (validationError) {
+      setValidationError(validationError)
+      showError('Failed to join party', validationError)
+      return
+    }
 
     router.push({
       pathname: '/lobby',
     })
+  }
 
-    setShowJoinModal(false)
+  const onChangePlayerName = (name: string) => {
+    setPlayerName(name)
+    setValidationError(undefined)
   }
 
   return (
-    <SafeAreaView className="flex-1 items-center bg-background-0 dark:bg-background-950">
-      <View className="flex-1 w-full max-w-md justify-evenly  p-4 items-center">
-        <View className="flex-col items-center gap-4 w-full">
-          <Text size="xl" style={{ width: 200, textAlign: 'center' }}>
-            Name
-          </Text>
-          <Input
-            variant="outline"
-            size="xl"
-            isDisabled={isLoadingPlayerName}
-            isInvalid={false}
-            isReadOnly={false}
-            style={{
-              alignItems: 'center',
-              justifyContent: 'center',
-              width: 200,
-            }}
-          >
-            <InputField
-              aria-label="Username"
-              value={playerName}
-              onChangeText={setPlayerName}
-              placeholder="Enter your name..."
-              autoComplete="username"
-              maxLength={PLAYER_NAME_MAX_LENGTH}
-              style={{ width: 200, textAlign: 'center' }}
-            />
-          </Input>
-        </View>
-        <View className="flex-row items-center gap-2">
-          <Heading size="4xl">
-            <RainbowText text="Party" className="text-6xl" />
-            <Text className="text-6xl"> Battle</Text>
-          </Heading>
-        </View>
-        <View className="flex-col w-full gap-4">
-          {partyCode && (
+    <>
+      <SafeAreaView className="flex-1 items-center bg-background-0 dark:bg-background-950">
+        <View className="flex-1 w-full max-w-md justify-evenly  p-4 items-center">
+          <View className="flex-col items-center gap-4 w-full">
+            <Text size="xl" style={{ width: 200, textAlign: 'center' }}>
+              Name
+            </Text>
+            <Input
+              variant="outline"
+              size="xl"
+              isDisabled={isLoadingPlayerName}
+              isInvalid={!!validationError}
+              style={{
+                alignItems: 'center',
+                justifyContent: 'center',
+                width: 200,
+              }}
+            >
+              <InputField
+                aria-label="Username"
+                value={playerName}
+                onChangeText={onChangePlayerName}
+                placeholder="Enter your name..."
+                autoComplete="username"
+                maxLength={PLAYER_NAME_MAX_LENGTH}
+                style={{ width: 200, textAlign: 'center' }}
+              />
+            </Input>
+          </View>
+          <View className="flex-row items-center gap-2">
+            <Heading size="4xl">
+              <RainbowText text="Party" className="text-6xl" />
+              <Text className="text-6xl"> Battle</Text>
+            </Heading>
+          </View>
+          <View className="flex-col w-full gap-4">
+            {partyCode && (
+              <View className="flex-row items-center justify-center w-full">
+                <Button
+                  size="xl"
+                  action="primary"
+                  variant="solid"
+                  onPress={() => handleJoinRoom(partyCode)}
+                  isDisabled={!trimmedPlayerName || isLoading}
+                  style={{ width: 200, paddingHorizontal: 8 }}
+                >
+                  <ButtonText>{isLoading ? 'Loading...' : `JOIN ${partyCode}`}</ButtonText>
+                </Button>
+              </View>
+            )}
+            <View className="flex-row items-center justify-center w-full">
+              <View className="flex-row gap-2">
+                <Button
+                  size="xl"
+                  action="primary"
+                  onPress={() => setShowJoinModal(true)}
+                  isDisabled={!trimmedPlayerName || isLoading}
+                  style={{ width: 200, paddingHorizontal: 8 }}
+                >
+                  <ButtonText>{isLoading ? 'Loading...' : 'JOIN PARTY'}</ButtonText>
+                </Button>
+              </View>
+            </View>
             <View className="flex-row items-center justify-center w-full">
               <Button
                 size="xl"
                 action="primary"
-                variant="solid"
-                onPress={() => handleJoinRoom(partyCode)}
+                variant="outline"
+                onPress={handleCreateRoom}
                 isDisabled={!trimmedPlayerName || isLoading}
                 style={{ width: 200, paddingHorizontal: 8 }}
               >
-                <ButtonText>{isLoading ? 'Loading...' : `JOIN ${partyCode}`}</ButtonText>
-              </Button>
-            </View>
-          )}
-          <View className="flex-row items-center justify-center w-full">
-            <View className="flex-row gap-2">
-              <Button
-                size="xl"
-                action="primary"
-                onPress={() => setShowJoinModal(true)}
-                isDisabled={!trimmedPlayerName || isLoading}
-                style={{ width: 200, paddingHorizontal: 8 }}
-              >
-                <ButtonText>{isLoading ? 'Loading...' : 'JOIN PARTY'}</ButtonText>
+                <ButtonText>{isLoading ? 'Loading...' : 'CREATE PARTY'}</ButtonText>
               </Button>
             </View>
           </View>
-          <View className="flex-row items-center justify-center w-full">
-            <Button
-              size="xl"
-              action="primary"
-              variant="outline"
-              onPress={handleCreateRoom}
-              isDisabled={!trimmedPlayerName || isLoading}
-              style={{ width: 200, paddingHorizontal: 8 }}
-            >
-              <ButtonText>{isLoading ? 'Loading...' : 'CREATE PARTY'}</ButtonText>
-            </Button>
+          <View className="h-10">
+            <StoreBadges />
           </View>
         </View>
-        <View className="h-10">
-          <StoreBadges />
-        </View>
-      </View>
+      </SafeAreaView>
 
       <JoinRoomModal
         isOpen={showJoinModal}
@@ -128,6 +143,6 @@ export default function HomeScreen() {
         onJoin={handleJoinRoom}
         isLoading={isLoading}
       />
-    </SafeAreaView>
+    </>
   )
 }